requirements.txt
numpy>=2.0.0
pandas>=2.0.0
pyyaml>=6.0
akshare>=1.10.0  # 数据源
vnpy>=3.0.0      # 交易框架(可选)